+++ src/components/content-layout/content-layout.js
@@ -1,16 +1,47 @@
 import { css } from 'linaria'
 import { cardsGrid } from 'components/items-layout/base'
+import { breakpointMediumTablet } from '@pocket/web-ui'
+import { breakpointSmallTablet } from '@pocket/web-ui'
+import { breakpointTinyTablet } from '@pocket/web-ui'
 
 export const contentLayout = css`
   .content-section {
     ${cardsGrid};
     position: unset;
+
+    ${breakpointSmallTablet} {
+      grid-column-gap: 16px;
+    }
   }
 
   header,
   footer,
   .content-body {
     grid-column: 2 / 9;
+
+    ${breakpointMediumTablet} {
+      grid-column: 2 / 10;
+    }
+
+    ${breakpointSmallTablet} {
+      grid-column: 2 / 11;
+    }
+
+    ${breakpointTinyTablet} {
+      grid-column: 1 / -1;
+    }
+
+    &.isMobileWebView {
+      grid-column: 1 / 8;
+
+      ${breakpointMediumTablet} {
+        grid-column: 1 / 12;
+      }
+
+      ${breakpointTinyTablet} {
+        grid-column: 1 / -1;
+      }
+    }
   }
 
   .left-aside {
@@ -20,6 +51,14 @@ export const contentLayout = css`
     & > div {
       flex-grow: 1;
     }
+
+    ${breakpointMediumTablet} {
+      position: static;
+      grid-column: 2 / -1;
+    }
+    ${breakpointTinyTablet} {
+      grid-column: 1 / -1;
+    }
   }
 
   .right-aside {
@@ -27,6 +66,10 @@ export const contentLayout = css`
     height: 100%;
     display: flex;
     flex-direction: column;
+    ${breakpointMediumTablet} {
+      grid-column: 0;
+      display: none;
+    }
     & > div {
       flex-grow: 1;
     }
@@ -39,14 +82,7 @@ export const contentLayout = css`
 
   .hero-image {
     width: 100%;
-    margin-bottom: 1rem;
-  }
-
-  .content-excerpt {
-    font-weight: 400;
-    font-size: 1.4375rem;
-    line-height: 160%;
-    margin-bottom: 2.5rem;
+    margin-bottom: 2rem;
   }
 
   article {