+++ src/components/content-saving/save-article.js
@@ -70,7 +70,7 @@ const saveArticleStyles = css`
 export function SaveArticleTop({ url, saveAction, saveStatus, isAuthenticated, trackSaveClick }) {
   const { t } = useTranslation()
   return (
-    <div className={classNames(saveArticleStyles, 'top')}>
+    <div className={classNames(saveArticleStyles, 'top', 'save-article')}>
       <SaveStoryButton
         trackSaveClick={trackSaveClick}
         isAuthenticated={isAuthenticated}
@@ -95,13 +95,7 @@ export function SaveArticleBottom({
   const { t } = useTranslation()
 
   return (
-    <aside className={classNames(saveArticleStyles, 'bottom')}>
-      <p>
-        {t(
-          'item-action:save-article-bottom',
-          'How was it? Save stories you love and never lose them.'
-        )}
-      </p>
+    <div className={classNames(saveArticleStyles, 'top', 'save-article')}>
       <SaveStoryButton
         trackSaveClick={trackSaveClick}
         isAuthenticated={isAuthenticated}
@@ -111,7 +105,13 @@ export function SaveArticleBottom({
         url={url}
         onClick={saveAction}
       />
-    </aside>
+      <p>
+        {t(
+          'item-action:save-article-bottom',
+          'How was it? Save stories you love and never lose them.'
+        )}
+      </p>
+    </div>
   )
 }
 
@@ -128,8 +128,8 @@ function SaveStoryButton({ url, isAuthenticated, saveStatus, saveAction, id, tra
     event.preventDefault()
     event.stopPropagation()
 
-    saveAction(url)
-    trackSaveClick(id)
+    saveAction(url, id)
+    if (trackSaveClick) trackSaveClick(id) // Legacy method: Remove on analytics confirmation
   }
 
   // Popover Effect