+++ src/components/global-nav/global-nav.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import Link from 'next/link'
 import PropTypes from 'prop-types'
 import { css } from 'linaria'
 import classnames from 'classnames'
@@ -95,6 +96,10 @@ const headerStyle = css`
     padding-top: 0;
     padding-bottom: 0;
     height: var(--size400);
+  }
+
+  @media print {
+    display: none;
   }
 `
 
@@ -259,17 +264,18 @@ const GlobalNav = ({
               isOpen={isMobileMenuOpen}
               toggleMenuOpen={setMobileMenuOpen}
             />
-            <a
-              id="pocket-logo-nav"
-              className="pocket-logo"
-              href={pocketLogoOutboundUrl}
-              onClick={(event) => {
-                handleLinkClick('pocket', event)
-              }}
-              data-cy="logo-link">
-              <Logo className="logo" />
-              {isLoggedIn ? <LogoMark className="logo-mark" /> : null}
-            </a>
+            <Link href={pocketLogoOutboundUrl} data-test="logo-link">
+              <a
+                id="pocket-logo-nav"
+                className="pocket-logo"
+                onClick={(event) => {
+                  handleLinkClick('pocket', event)
+                }}
+                data-cy="logo-link">
+                <Logo className="logo" />
+                {isLoggedIn ? <LogoMark className="logo-mark" /> : null}
+              </a>
+            </Link>
           </div>
           {children ? (
             children