+++ src/containers/collections/collection-page.js
@@ -1,9 +1,11 @@
+import { BASE_URL } from 'common/constants'
 import Layout from 'layouts/main'
 import MobileLayout from 'layouts/mobile-web'
 
 import { useDispatch, useSelector } from 'react-redux'
 
 import { contentLayout } from 'components/content-layout/content-layout'
+import { printLayout } from 'components/content-layout/print-layout'
 import { PocketWorthy } from 'components/content-headline/pocket-worthy'
 import { ParsedHeadline } from 'components/content-headline/parsed-headline'
 import { AuthorByline } from 'components/content-author/author-byline'
@@ -14,56 +16,74 @@ import { AdAboveTheFold } from 'components/content-ads/content-ads'
 import { AdBelowTheFold } from 'components/content-ads/content-ads'
 import { AdRailTop } from 'components/content-ads/content-ads'
 import { AdRailBottom } from 'components/content-ads/content-ads'
+import { ContentIntro } from 'components/content-intro/content-intro'
+import { AuthorBio } from 'components/content-author/author-bio'
 
+import { getImageCacheUrl } from 'common/utilities'
 import { CardTopicsNav as TopicsBubbles } from 'connectors/topic-list/topic-list'
 import { ItemCard } from 'connectors/item-card/collection/story-card'
 import { saveCollection } from 'containers/collections/collections.state'
+
+import { unSaveCollectionPage } from 'containers/collections/collections.state'
 import { saveCollectionPage } from 'containers/collections/collections.state'
+import { trackItemSave } from 'connectors/snowplow/snowplow.state'
+
 import { Toasts } from 'connectors/toasts/toast-list'
-import { useTranslation } from 'next-i18next'
+import ErrorPage from 'pages/_error'
 
-export function CollectionPage({ queryParams = {}, slug }) {
-  const { t } = useTranslation()
-
-  const metaData = {
-    description: t(
-      'discover:page-description',
-      'Discover fascinating stories from all across the web with Pocket.'
-    ),
-    title: t('discover:page-title', 'Discover stories on Pocket')
-  }
+export function CollectionPage({ queryParams = {}, slug, statusCode }) {
   const dispatch = useDispatch()
-
-  const isAuthenticated = useSelector((state) => state.user.auth)
-  const data = useSelector((state) => state.collections[slug]) || {}
-  const topics = useSelector((state) => state.topicList?.topicsByName)
 
   const { mobile_web_view: isMobileWebView } = queryParams
   const ArticleLayout = isMobileWebView ? MobileLayout : Layout
 
-  const { title, intro, excerpt, authors, stories, imageUrl, urls, pageSaveStatus } = data
+  const isAuthenticated = useSelector((state) => state.user?.auth)
+  const isPremium = useSelector((state) => state.user?.premium_status)
+  const oneTrustReady = useSelector((state) => state.oneTrust?.trustReady)
+  const trackingEnabled = useSelector((state) => state.oneTrust?.tracking?.enabled)
+  const data = useSelector((state) => state.collections[slug]) || {}
+  const topics = useSelector((state) => state.topicList?.topicsByName)
+  const userStatus = useSelector((state) => state.user.user_status)
+  const shouldRender = userStatus !== 'pending'
+
+  // Show error page if things have gone awry
+  if (statusCode) return <ErrorPage statusCode={statusCode} />
+
+  const { title, intro, excerpt, authors, stories, imageUrl, pageSaveStatus } = data
+  const { showAds = true } = data
   const authorNames = authors?.map((author) => author.name)
+  const allowAds = isPremium ? false : showAds && shouldRender && oneTrustReady
+  const usePersonalized = allowAds && trackingEnabled
+  const heroImage = getImageCacheUrl(imageUrl, { width: 648 })
 
-  const saveAction = () => dispatch(saveCollectionPage(slug))
+  // const count = urls?.length
+  // const saveCollectionTop = () => dispatch(saveCollection(slug))
+  // const saveCollectionBottom = () => dispatch(saveCollection(slug))
+  const url = `${BASE_URL}/collections/${slug}`
+  const metaData = { description: excerpt, title, url, image: imageUrl }
 
-  const count = urls.length
-  const saveCollectionTop = () => dispatch(saveCollection(slug))
-  const saveCollectionBottom = () => dispatch(saveCollection(slug))
+  const saveAction = (saveUrl, id) => {
+    if (pageSaveStatus === 'saved') dispatch(unSaveCollectionPage(slug))
+    if (pageSaveStatus !== 'saved') {
+      dispatch(saveCollectionPage(slug))
+      dispatch(trackItemSave(0, { url }, id))
+    }
+  }
 
   return (
-    <ArticleLayout title={metaData.title} metaData={metaData}>
+    <ArticleLayout title={metaData.title} metaData={metaData} className={printLayout}>
       <main className={contentLayout}>
         <section>
-          <AdAboveTheFold />
+          <AdAboveTheFold allowAds={allowAds} usePersonalized={usePersonalized} />
         </section>
         {/* Content header information */}
         <section className="content-section">
           <header>
             <PocketWorthy />
-            <ParsedHeadline title={title} description={intro} />
+            <ParsedHeadline title={title} description={excerpt} useMarkdown={true} />
             {authors ? (
               <AuthorByline
-                url="/collections"
+                url="/explore"
                 name="Pocket Collections"
                 showAuthors={true}
                 authorNames={authorNames}
@@ -85,7 +105,7 @@ export function CollectionPage({ queryParams = {}, slug }) {
               isMobileWebView={isMobileWebView}
               title={title}
               excerpt={excerpt}
-              saveAction={saveAction}
+              onSave={saveAction}
               saveStatus={pageSaveStatus}
               isAuthenticated={isAuthenticated}
               handleShareClick={() => {}}
@@ -96,18 +116,23 @@ export function CollectionPage({ queryParams = {}, slug }) {
 
           {/* Right aside content such as ads and recs */}
           <aside className="right-aside">
-            <AdRailTop />
-            <AdRailBottom />
+            <AdRailTop allowAds={allowAds} usePersonalized={usePersonalized} />
+            <AdRailBottom allowAds={allowAds} usePersonalized={usePersonalized} />
           </aside>
 
           <div className="content-body">
-            <img src={imageUrl} alt="" className="hero-image" />
-            <p className="content-excerpt">{excerpt}</p>
+            <img src={heroImage} alt="" className="hero-image" />
+
+            <ContentIntro intro={intro} />
 
             {/* Collection Stories */}
-            {stories.map((id, index) => (
-              <ItemCard id={id} key={id} position={index} cardShape="wide" showExcerpt={true} />
-            ))}
+            {stories
+              ? stories.map((id, index) => (
+                  <ItemCard id={id} key={id} position={index} cardShape="wide" showExcerpt={true} />
+                ))
+              : null}
+
+            {authors ? authors?.map((author) => <AuthorBio {...author} />) : null}
           </div>
         </section>
 
@@ -119,7 +144,7 @@ export function CollectionPage({ queryParams = {}, slug }) {
               saveStatus={pageSaveStatus}
             />
 
-            <AdBelowTheFold />
+            <AdBelowTheFold allowAds={allowAds} usePersonalized={usePersonalized} />
 
             <TopicsBubbles topics={topics} className="no-border" />
           </footer>