+++ src/pages/_app.js
@@ -5,7 +5,6 @@ import { appWithTranslation } from 'next-i18next'
 import { useEffect } from 'react'
 import { wrapper } from 'store'
 import { useDispatch, useSelector } from 'react-redux'
-import { useRouter } from 'next/router'
 import * as Sentry from '@sentry/node'
 import { parseCookies } from 'nookies'
 
@@ -15,22 +14,18 @@ import { getSessGuid, sessGuidHydrate } from 'connectors/user/user.state'
 import { fetchUnleashData } from 'connectors/feature-flags/feature-flags.state'
 import { featuresHydrate } from 'connectors/feature-flags/feature-flags.state'
 
-import { hydrateUserTags } from 'containers/my-list/tags-page/tags-page.state'
-
 import { appSetPreferences } from 'connectors/app/app.state'
+import { hydrateSettings } from 'connectors/settings/settings.state'
 
 import { legacyAnalyticsTrack } from 'common/api/legacy-analytics'
+
+import { PostTrustInit } from 'connectors/one-trust/one-trust'
 
 /** Setup Files
  --------------------------------------------------------------- */
 import { sentrySettings } from 'common/setup/sentry'
 import { loadPolyfills } from 'common/setup/polyfills'
-import { initializeSnowplow } from 'common/setup/snowplow'
 import { signalTestsReady } from '../../cypress/support/utils'
-
-import { trackPageView } from 'connectors/snowplow/snowplow.state'
-import { GOOGLE_ANALYTICS_ID } from 'common/constants'
-import ReactGA from 'react-ga'
 
 import { Shortcuts } from 'connectors/shortcuts/shortcuts'
 import { DevTools } from 'connectors/dev-tools/dev-tools'
@@ -44,13 +39,9 @@ Sentry.init(sentrySettings)
 function PocketWebClient({ Component, pageProps, err }) {
   // Initialize app once per page load
   const dispatch = useDispatch()
-  const router = useRouter()
 
   const { user_status, user_id, sess_guid, birth } = useSelector((state) => state.user) //prettier-ignore
-  const path = router.pathname
-
   const showDevTools = process.env.SHOW_DEV === 'included'
-
   const { authRequired } = pageProps
 
   useEffect(() => {
@@ -75,6 +66,7 @@ function PocketWebClient({ Component, pageProps, err }) {
 
     // Set up defaults/user pref in state
     dispatch(appSetPreferences())
+    dispatch(hydrateSettings())
 
     /**
      * First time user
@@ -103,22 +95,6 @@ function PocketWebClient({ Component, pageProps, err }) {
     if (sess_guid) validateUser()
   }, [user_status, dispatch])
 
-  // 3rd party initializations
-  useEffect(() => {
-    if (user_status === 'pending' || user_status === 'invalid' || !sess_guid) {
-      return
-    }
-
-    // Set up Snowplow
-    initializeSnowplow(user_id, sess_guid)
-
-    // Set up Google Analytics
-    ReactGA.initialize(GOOGLE_ANALYTICS_ID)
-
-    // Load OptinMonster
-    // loadOptinMonster()
-  }, [user_id, user_status, sess_guid])
-
   // Hydrate user features
   useEffect(() => {
     if (user_status === 'pending' || user_status === 'invalid') return null
@@ -130,21 +106,10 @@ function PocketWebClient({ Component, pageProps, err }) {
 
     hydrateFeatures()
 
-    // Hydrate locally stored values
-    dispatch(hydrateUserTags())
-
     // signal to Cypress that React client side has loaded
     // Make sure this is the last thing we fire
     signalTestsReady()
-  }, [user_status, sess_guid, user_id, dispatch])
-
-  // Track Page Views
-  useEffect(() => {
-    if (user_status === 'pending' || user_status === 'invalid') return null
-
-    dispatch(trackPageView())
-    ReactGA.pageview(path)
-  }, [user_status, sess_guid, user_id, path, dispatch])
+  }, [user_status, sess_guid, user_id, birth, dispatch])
 
   useEffect(() => {
     if (authRequired && user_status === 'invalid') {
@@ -153,12 +118,11 @@ function PocketWebClient({ Component, pageProps, err }) {
   }, [authRequired, user_status])
 
   // Provider is created automatically by the wrapper by next-redux-wrapper
-  const shouldRender = authRequired
-    ? user_status !== 'pending' && user_status !== 'invalid'
-    : true
+  const shouldRender = authRequired ? user_status !== 'pending' && user_status !== 'invalid' : true
 
   return (
     <ViewportProvider>
+      <PostTrustInit />
       {showDevTools ? <DevTools /> : null}
       <Shortcuts />
       {shouldRender ? <Component {...pageProps} err={err} /> : null}